feat: refatora repositorio cliente
