--- src/repositorio/cliente/interfaces/compostas/ClienteInativo.java
@@ -1,7 +1,0 @@
-package repositorio.cliente.interfaces.compostas;
-
-import repositorio.cliente.interfaces.especificas.ClienteManipulacao;
-import repositorio.cliente.interfaces.especificas.ClienteConsulta;
-
-public interface ClienteInativo extends ClienteConsulta, ClienteManipulacao {
-}
