+++ src/repositorio/cliente/interfaces/compostas/RepositorioCliente.java
@@ -2,8 +2,7 @@ package repositorio.cliente.interfaces.compostas;
 
 import repositorio.cliente.interfaces.especificas.ClienteConsulta;
 import repositorio.cliente.interfaces.especificas.ClienteManipulacao;
-import repositorio.cliente.interfaces.especificas.ClienteAlteracao;
 
-public interface ClienteAtivo extends ClienteConsulta, ClienteManipulacao, ClienteAlteracao {
+public interface RepositorioCliente extends ClienteConsulta, ClienteManipulacao {
 
 }
