--- src/repositorio/cliente/interfaces/especificas/ClienteAlteracao.java
@@ -1,7 +1,0 @@
-package repositorio.cliente.interfaces.especificas;
-
-import dominio.Cliente.Cliente;
-
-public interface ClienteAlteracao {
-    void alterarEmail(Cliente cliente, String novoEmail);
-}
