+++ src/repositorio/cliente/interfaces/especificas/ClienteConsulta.java
@@ -5,7 +5,9 @@ import dominio.Cliente.Cliente;
 import java.util.List;
 
 public interface ClienteConsulta {
-    List<Cliente> listarClientes();
+    List<Cliente> listarClientesAtivos();
+    List<Cliente> listarClientesInativos();
+    List<Cliente> listarTodosClientes();
     Cliente buscarPorDocumento(String documento);
     Cliente buscarPorId(long id);
 }
